Add files via upload
This workbook includes an attempt at implementing the H2 ground state estimator as described in the paper arXiv:1611.01068v1 [quant-ph] 3 Nov 2016 . The attempt is currently a fail in that the ground state calculated do not match the targets as per the data table in the document. This should be reviewed but due to the amount of anticipated efforts it should be best left for after we have achieved more pressing goals.